add `RouteAddRequest<IpAddr>`
We currently have `RouteAddRequest<Ipv4Addr>` and
`RouteAddRequest<Ipv6Addr>`. But when working on codebases that mostly
deal with `IpAddr`, this is an inconvenient API because we have to
match on every address, and are forced to write the same code twice.

The new impl makes working with `IpAddr` much more convenient, with
the downside that the requests are validated at runtime instead of
compile time.