Replace `impl TryStream` with `impl Stream`
The problem with `impl TryStream` is that the opaque type cannot be used as `Stream` (thus `StreamExt`) anymore, only `TryStream`, because types implementing the latter does not necessarily implement the former. This also causes problems with dynamic dispatch. By making this change we can use methods from both traits and should not break users' current usage.